feat: implement subscribeToRoom method #69
