+++ src/main/java/com/kospot/infrastructure/exception/payload/code/ErrorStatus.java
@@ -80,6 +80,7 @@ public enum ErrorStatus implements BaseCode {
     GAME_ROOM_MEMBER_ALREADY_IN_ROOM(BAD_REQUEST, 4306, "이미 게임 방에 참여 중입니다."),
     GAME_ROOM_IS_NOT_ENOUGH_PLAYER(BAD_REQUEST, 4307, "게임 방에 참여할 플레이어가 부족합니다."),
     GAME_ROOM_CANNOT_JOIN_NOW(BAD_REQUEST, 4308, "현재 게임 방에 참여할 수 없습니다."),
+    GAME_ROOM_PLAYER_NOT_FOUND(NOT_FOUND, 4309, "해당 게임 방 플레이어를 찾을 수 없습니다."),
 
     // MultiGame Error(4311 ~ 4320)
     PLAYER_MATCH_TYPE_NOT_FOUND(NOT_FOUND, 4311, "해당 플레이어 매치 타입을 찾을 수 없습니다."),
