+++ src/main/java/com/kospot/presentation/multigame/gameroom/controller/GameRoomWebSocketController.java
@@ -1,9 +1,12 @@
 package com.kospot.presentation.multigame.gameroom.controller;
 
+import com.kospot.application.multiplayer.gameroom.websocket.usecase.JoinGameRoomSocketUseCase;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
+import org.springframework.messaging.simp.SimpMessageHeaderAccessor;
 import org.springframework.stereotype.Controller;
 
 @Slf4j
@@ -11,6 +14,8 @@ import org.springframework.stereotype.Controller;
 @RequiredArgsConstructor
 @Tag(name = "GameRoom WebSocket", description = "게임방 실시간 통신")
 public class GameRoomWebSocketController {
+
+    private final JoinGameRoomSocketUseCase joinGameRoomSocketUseCase;
 
     /**
      * 게임방 채팅 메시지 전송
@@ -28,8 +33,10 @@ public class GameRoomWebSocketController {
      * 구독: /topic/room/123/playerList, /topic/room/123/status
      */
     @MessageMapping("/room.{roomId}.subscribe")
-    public void subscribeToRoom(){
+    public void subscribeToRoom(@DestinationVariable String roomId,
+                                SimpMessageHeaderAccessor headerAccessor){
 
+        joinGameRoomSocketUseCase.execute(roomId, headerAccessor);
     }
 
 }
