+++ src/test/java/com/kospot/application/multiplayer/gameroom/usecase/JoinGameRoomConcurrencyTest.java
@@ -1,5 +1,6 @@
 package com.kospot.application.multiplayer.gameroom.usecase;
 
+import com.kospot.application.multiplayer.gameroom.http.usecase.JoinGameRoomUseCase;
 import com.kospot.domain.game.vo.GameMode;
 import com.kospot.domain.member.entity.Member;
 import com.kospot.domain.member.repository.MemberRepository;
